+++ shared/libebm/BoosterCore.cpp
@@ -1045,7 +1045,7 @@ ErrorEbm BoosterCore::InitializeBoosterGradientsAndHessians(
 
          ApplyUpdateBridge data;
          data.m_cScores = cScores;
-         data.m_cPack = static_cast<int>(k_cItemsPerBitPackNone);
+         data.m_cPack = k_cItemsPerBitPackNone;
          data.m_bHessianNeeded = IsHessian() ? EBM_TRUE : EBM_FALSE;
          data.m_bValidation = EBM_FALSE;
          data.m_aMulticlassMidwayTemp = aMulticlassMidwayTemp;
